Fix bugs in Payment and FieldSchedule

+++ Mini-soccer-field-manager/DTO_MANAGER/drinksDTO.cs
@@ -13,15 +13,15 @@ namespace DTO_MANAGER
         private String drinksID;
         private String drinksName;
         private int drinksLeft;
-        private int drinksPurchasePrice;
-        private int drinksSalePrice;
+        private float drinksPurchasePrice;
+        private float drinksSalePrice;
         private DateTime billDate;
 
         public String MaDoUong { get => drinksID; set => drinksID = value; }
         public String TenDoUong { get => drinksName; set => drinksName = value; }
         public int DoUongConLai { get => drinksLeft; set => drinksLeft = value; }
-        public int DonGiaMua { get => drinksPurchasePrice; set => drinksPurchasePrice = value; }
-        public int DonGiaBan { get => drinksSalePrice; set => drinksSalePrice = value; }
+        public float DonGiaMua { get => drinksPurchasePrice; set => drinksPurchasePrice = value; }
+        public float DonGiaBan { get => drinksSalePrice; set => drinksSalePrice = value; }
         public DateTime NgayHoaDon { get => billDate; set => billDate = value; }
     }
 }
